Add intersection to enumerable utils
